Minor modifications. Tried to implement timer for blinking, problems with variables scopes.

+++ moka_twi.h
@@ -23,10 +23,12 @@
 
 #include <Arduino.h>
 
+#include "moka_leds.h"
+#include "moka_pad.h"
 
 void mw_init();
 
-void mw_receiveHandler(uint8_t bytes);
+void mw_receiveHandler(int bytes);
 void mw_requestHandler();
